# backend/api/bundles/backend.openapi.v1.bundle.yaml
openapi: 3.0.3
info:
  title: Backend API
  version: 1.0.0
  description: Backend для работы анализатора продуктов
  license:
    url: https://github.com/arseniizyk/food-analyser/blob/07936fac66da411a4cb4464d4cb4e53e7f857cec/LICENSE
    name: MIT
servers:
  - url: http://0.0.0.0:8000
    description: Local Development
  - url: http://food_backend:8000
    description: Inside the docker container
security: []
tags:
  - name: Backend
    description: операции с данными о продуктах и юзерах
paths:
  /api/v1/health:
    get:
      summary: Healthcheck
      operationId: health
      responses:
        '200':
          description: Service is healthy
        '400':
          description: Bad request
        '503':
          description: Service is unhealthy
  /api/v1/analysis/{barcode}:
    parameters:
      - $ref: '#/components/parameters/barcode'
    get:
      summary: Get analysis by barcode
      operationId: GetAnalysisByBarcode
      tags:
        - Barcode, Analysis
      responses:
        '200':
          description: Analysis information successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/analysis'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Barcode analysis wasn't found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/analyze/{barcode}:
    parameters:
      - $ref: '#/components/parameters/barcode'
    post:
      summary: Analyze analysis nutrition photo
      operationId: Analyze
      tags:
        - Barcode, Analysis
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                user_id:
                  type: string
                  description: Optional user ID (if authenticated). If not provided, analysis is stored only locally on client.
      responses:
        '200':
          description: Analyze has competed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/analysis'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        '502':
          description: LLM Service isn't responding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_gateway_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/auth/google:
    post:
      tags:
        - Authentication
      summary: Authenticate with Google
      operationId: authenticateWithGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_google_request'
      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth_response'
        '401':
          description: Invalid Google token.
        '500':
          description: Internal server error.
  /api/v1/history:
    get:
      summary: Get user scans
      operationId: getUserScans
      tags:
        - Analysis
      responses:
        '200':
          description: User's scans found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user_scans_history'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
components:
  parameters:
    barcode:
      name: barcode
      in: path
      required: true
      description: Barcode продукта, для которого запрашивается анализ
      schema:
        type: string
        minLength: 1
        maxLength: 15
        example: '5901234123457'
  schemas:
    generic_error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
        message:
          type: string
          description: Описание ошибки
    summary_item:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Composition looks balanced.
    risk:
      type: object
      required:
        - severity
        - title
        - description
      properties:
        severity:
          type: string
          enum:
            - low
            - medium
            - high
        title:
          type: string
          example: High sugar
        description:
          type: string
          example: Contains a significant amount of added sugar.
    ingredient:
      type: object
      required:
        - name
        - risk
      properties:
        name:
          type: string
          example: Sugar
        risk:
          type: string
          enum:
            - safe
            - caution
            - dangerous
        description:
          type: string
          nullable: true
    analysis:
      type: object
      required:
        - barcode
        - score
        - grade
        - summary
        - risks
        - ingredients
      properties:
        barcode:
          type: string
          example: '978020137962'
        score:
          type: integer
          description: Оценка продукта
          minimum: 0
          maximum: 100
          example: 85
        grade:
          type: string
          enum:
            - excellent
            - good
            - average
            - poor
        summary:
          type: array
          items:
            $ref: '#/components/schemas/summary_item'
        risks:
          type: array
          items:
            $ref: '#/components/schemas/risk'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/ingredient'
    bad_request_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 400
        message:
          type: string
          description: Описание ошибки
          example: 'Bad Request: Invalid parameter format'
    not_found_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Описание ошибки
          example: Analysis for barcode not found
    internal_server_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 500
        message:
          type: string
          description: Описание ошибки
          example: Internal server error occurred
    bad_gateway_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 502
        message:
          type: string
          description: Описание ошибки
          example: LLM Service isn't responding
    auth_google_request:
      type: object
      required:
        - id_token
      properties:
        id_token:
          type: string
    auth_response:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          description: Unique user ID created or retrieved from backend
    user_scans_history:
      type: array
      items:
        type: object
        description: Barcode
        required:
          - barcode
          - created_at
          - score
        properties:
          barcode:
            type: string
            example: '4601234567890'
          created_at:
            type: string
            format: date-time
            example: '2026-08-15T20:50:00Z'
          score:
            type: integer
            example: 78